requirements.txt
numpy~=2.2
pillow~=11.1
